Rename VersionTag.NS_VTAG to VersionTag.NAMESPACE
git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/projects/lily/trunk@4024 b80d95f8-e103-0410-bd55-e03a9dc29394

+++ repo-util/src/main/java/org/lilycms/repoutil/RecordEvent.java
@@ -47,11 +47,11 @@ public class RecordEvent {
         JsonNode msgData = new ObjectMapper().readValue(data, 0, data.length, JsonNode.class);
 
         if (msgData.get("versionCreated") != null) {
-            versionCreated = msgData.get("versionCreated").getIntValue();
+            versionCreated = msgData.get("versionCreated").getLongValue();
         }
 
         if (msgData.get("versionUpdated") != null) {
-            versionUpdated = msgData.get("versionUpdated").getIntValue();
+            versionUpdated = msgData.get("versionUpdated").getLongValue();
         }
 
         if (msgData.get("recordTypeChanged") != null) {
